==> notebooks/1.0-pvso-initial-exploration.ipynb <==
# Importação

import pandas as pd

# Lendo os dataframes 2010-2022

df_2010_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20101.csv')
df_2010_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20102.csv')

df_2011_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20111.csv')
df_2011_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20112.csv')

df_2012_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20121.csv')
df_2012_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20122.csv')

df_2013_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20131.csv')
df_2013_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20131.csv')

df_2014_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20141.csv')
df_2014_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20142.csv')

df_2015_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20151.csv')
df_2015_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20152.csv')

df_2016_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20161.csv')
df_2016_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20162.csv')

df_2017_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20171.csv')
df_2017_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20172.csv')

df_2018_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20181.csv')
df_2018_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20182.csv')

df_2019_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20191.csv')
df_2019_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20192.csv')

df_2020_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20201.csv')
df_2020_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20202.csv')

df_2021_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20211.csv')
df_2021_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20212.csv')

df_2022_1 = pd.read_csv('../data/raw/emprestimos/emprestimos-20221.csv')
df_2022_2 = pd.read_csv('../data/raw/emprestimos/emprestimos-20222.csv')

df_2020_2022 = pd.concat([
    df_2020_2, df_2021_1, df_2021_2, df_2022_1, df_2022_2
    ], ignore_index=True)

df_2020_2022['tipo_vinculo_usuario'].unique()

df_2020_2022.columns

tipo_vinculo_usuario = {
    'ALUNO DE PÃ“S-GRADUAÃ‡ÃƒO': 'ALUNO DE PÓS-GRADUAÇÃO',
    'DOCENTE': 'DOCENTE',
    'ALUNO DE GRADUAÃ‡ÃƒO': 'ALUNO DE GRADUAÇÃO',
    'SERVIDOR TÃ‰CNICO-ADMINISTRATIVO': 'SERVIDOR TÉCNICO-ADMINISTRATIVO',
    'ALUNO MÃ‰DIO/TÃ‰CNICO': 'ALUNO MÉDIO/TÉCNICO',
    'DOCENTE EXTERNO': 'DOCENTE EXTERNO',
    'USUÃ\x81RIO EXTERNO': 'USUÁRIO EXTERNO'
}

df_2020_2022['tipo_vinculo_usuario'] = df_2020_2022['tipo_vinculo_usuario'].map(tipo_vinculo_usuario)


df_2020_2022['tipo_vinculo_usuario'].unique()

df_2020_2022.drop(columns='nome_usuario', inplace=True)

df_2020_2022.columns

# Concatenando todos os dataframes em um

df_concat = pd.concat([
    df_2010_1, df_2010_2,
    df_2011_1, df_2011_2,
    df_2012_1, df_2012_2,
    df_2013_1, df_2013_2,
    df_2014_1, df_2014_2,
    df_2015_1, df_2015_2,
    df_2016_1, df_2016_2,
    df_2017_1, df_2017_2,
    df_2018_1, df_2018_2,
    df_2019_1, df_2019_2,
    df_2020_1, df_2020_2022
], ignore_index=True)

df_concat['tipo_vinculo_usuario'].unique()

# Mostrando alguns registros da concatenação
df_concat.head()

df_concat.tail()

# Mostrando os tipos dos dados da base de dados
df_concat.info()

# Mostrando o tamanho da base de dados (linhas e colunas)
df_concat.shape

# Mostrando as colunas da base de dados
df_concat.columns

# Verificando o tipo de dado de cada coluna
df_concat.dtypes

# Significado de cada coluna:

- id_emprestimo - Identificador do empréstimo
- codigo_barras - Código de barras do item do acervo
- data_renovacao - Data da renovação do empréstimo.
- data_emprestimo - Data da realização do empréstimo.
- data_devolucao - Data da devolução do item
- matricula_ou_siape - Matricula ou siape do usuário que realizou o
empréstimo.
- tipo_vinculo_usuario - Tipo do vínculo do usuário (servidor, discente,
docente etc.). 


# Verificando se há registros nulos
df_concat.isnull().sum()

# Verificando se há duplicatas
df_concat.value_counts()

df_concat.drop_duplicates(inplace=True)

# Verificando se há duplicatas
df_concat.value_counts()

# Resumo estatístico da base de dados com dados categoricos
df_concat.describe(include='object')

df_concat.to_csv('../data/interim/emprestimos-concat.csv', index=False)

# Lendo o arquivo que contém os exemplares
df_exemplares = pd.read_parquet('../data/raw/dados_exemplares.parquet')

# Mostrando alguns registros dos exemplares
df_exemplares.head()

# Mostrando os tipos dos dados dos exemplares
df_exemplares.info()

# Tamanho do dataframe exemplares
df_exemplares.shape

# Mostrando as colunas do dataframe exemplares
df_exemplares.columns

# Verificando o tipo de dado de cada coluna do dataframe exemplares
df_exemplares.dtypes

# Verificando se há registros nulos no dataframe exemplares
df_exemplares.isnull().sum()

# Resumo estatístico do dataframe exemplares com dados categoricos
df_exemplares.describe(include='object')
